requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
asgiref==3.8.1
    # via django
astroid==3.3.8
    # via pylint
colorama==0.4.6
    # via pylint
dill==0.3.9
    # via pylint
django==5.1.2
    # via
    #   -r requirements.in
    #   djangorestframework
djangorestframework==3.15.2
    # via -r requirements.in
isort==5.13.2
    # via pylint
mccabe==0.7.0
    # via pylint
platformdirs==4.3.6
    # via pylint
pylint==3.3.3
    # via -r requirements.in
sqlparse==0.5.2
    # via django
tomlkit==0.13.2
    # via pylint
tzdata==2024.2
    # via django
